requirements.txt
requests
python-dotenv
SQLAlchemy>=2.0
psycopg2-binary
google-generativeai


